Tests e2e playwright: gotoMap global method
