+++ tests/end2end/playwright/base-layers.spec.js
@@ -1,21 +1,6 @@
 // @ts-check
 const { test, expect } = require('@playwright/test');
-
-const gotoMap = async (url, page) => {
-    // Wait for WMS GetCapabilities promise
-    let getCapabilitiesWMSPromise = page.waitForRequest(/SERVICE=WMS&REQUEST=GetCapabilities/);
-    // Wait for WMS GetLegendGraphic promise
-    const getLegendGraphicPromise = page.waitForRequest(request => request.method() === 'POST' && request.postData() != null && request.postData()?.includes('GetLegendGraphic') === true);
-    await page.goto(url);
-    // Wait for WMS GetCapabilities
-    await getCapabilitiesWMSPromise;
-    // Wait for WMS GetLegendGraphic
-    await getLegendGraphicPromise;
-    // Check no error message displayed
-    expect(page.getByText('An error occurred while loading this map. Some necessary resources may temporari')).toHaveCount(0);
-    // Wait to be sure the map is ready
-    await page.waitForTimeout(1000)
-}
+const { gotoMap } = require('./globals')
 
 test.describe('Base layers', () => {
